requirements.txt
requests
notion-client
github-heatmap
retrying
pendulum
python-dotenv
pycryptodome